Import Protocol JS components as ES modules
