+++ bedrock/base/templates/base-article.html
@@ -49,6 +49,7 @@
 {% block site_js %}
   {{ super() }}
   <!--[if !IE]><!-->
+    {{ js_bundle('protocol-sidemenu') }}
     {{ js_bundle('basic-article') }}
     {{ js_bundle('newsletter') }}
   <!--<![endif]-->
