+++ bedrock/careers/templates/careers/home.html
@@ -414,6 +414,7 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('protocol-modal') }}
   {{ js_bundle('careers-hero-video') }}
   {{ js_bundle('careers-testimonials') }}
 {% endblock %}
