+++ bedrock/firefox/templates/firefox/all-unified.html
@@ -353,5 +353,7 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('protocol-modal') }}
+  {{ js_bundle('protocol-sidemenu') }}
   {{ js_bundle('firefox_all_unified') }}
 {% endblock %}
