+++ bedrock/firefox/templates/firefox/welcome/page6.html
@@ -120,5 +120,6 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('protocol-modal') }}
   {{ js_bundle('firefox_welcome_page6') }}
 {% endblock %}
