+++ bedrock/mozorg/templates/mozorg/about/leadership/base.html
@@ -15,10 +15,6 @@
   {{ css_bundle('about-leadership') }}
 {% endblock %}
 
-{% block js %}
-  {{ js_bundle('about-leadership') }}
-{% endblock %}
-
 {% set navigation_bar = [
   (url('mozorg.about.leadership.index'), 'executive-team', 'Leadership'),
   (url('mozorg.about.leadership.index'), 'executive-team', 'Executive Team'),
@@ -43,4 +39,10 @@
   </main>
 </div>
 <div class="mzp-u-modal-content"></div>
+{% endblock %}
+
+{% block js %}
+  {{ js_bundle('protocol-modal') }}
+  {{ js_bundle('protocol-sidemenu') }}
+  {{ js_bundle('about-leadership') }}
 {% endblock %}
