+++ bedrock/mozorg/templates/mozorg/about/webvision/summary.html
@@ -14,11 +14,6 @@
   {{ css_bundle("webvision") }}
 {% endblock %}
 
-{% block js %}
-  {{ js_bundle('webvision') }}
-{% endblock %}
-
-
 {% block content %}
 <main class="c-webvision-article-summary mzp-l-content">
   <article class="mzp-c-article c-webvision-article">
@@ -34,4 +29,9 @@
     <p class="mzp-c-call-out-desc">This was just the summary. You can read our <a href="{{ url('mozorg.about.webvision.full') }}">entire vision for the evolution of the Web here</a>.</p>
   </div>
 </section>
+{% endblock %}
+
+{% block js %}
+  {{ js_bundle('protocol-sidemenu') }}
+  {{ js_bundle('basic-article') }}
 {% endblock %}
