+++ bedrock/mozorg/templates/mozorg/contact/contact-base.html
@@ -72,6 +72,6 @@
 {% block js %}
   {# Issue 9322 #}
   {% if self.body_id() != 'contact-landing' %}
-    {{ js_bundle('contact-spaces') }}
+    {{ js_bundle('protocol-sidemenu') }}
   {% endif %}
 {% endblock %}
