+++ bedrock/mozorg/templates/mozorg/diversity/2021/index.html
@@ -19,10 +19,6 @@
   {{ css_bundle('diversity-2021') }}
 {% endblock %}
 
-{% block js %}
-  {{ js_bundle('diversity-2021') }}
-{% endblock %}
-
 {% block sub_navigation %}
 {{ sub_nav(
   title={
@@ -222,4 +218,9 @@
 </main>
 
 <div class="mzp-u-modal-content"></div>
+{% endblock %}
+
+{% block js %}
+  {{ js_bundle('protocol-modal') }}
+  {{ js_bundle('diversity-2021') }}
 {% endblock %}
