+++ bedrock/mozorg/templates/mozorg/home/home-contentful.html
@@ -122,6 +122,7 @@
 
 {% block js %}
   {% include 'includes/contentful/js.html' %}
+  {{ js_bundle('protocol-modal') }}
   {{ js_bundle('home-contentful') }}
   {{ js_bundle('newsletter') }}
