+++ bedrock/mozorg/templates/mozorg/home/home-de.html
@@ -164,6 +164,7 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('protocol-modal') }}
   {{ js_bundle('home') }}
   {{ js_bundle('newsletter') }}
