+++ bedrock/mozorg/templates/mozorg/home/home-fr.html
@@ -163,6 +163,7 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('protocol-modal') }}
   {{ js_bundle('home') }}
   {{ js_bundle('newsletter') }}
