+++ bedrock/privacy/templates/privacy/archive/firefox-os-2022-05.html
@@ -44,5 +44,6 @@
 {% block footnote %}<p><a href="{{ url('privacy') }}#contact">Contact Us</a></p>{% endblock %}
 
 {% block js %}
+  {{ js_bundle('protocol-sidemenu') }}
   {{ js_bundle('privacy_protocol') }}
 {% endblock %}
