+++ bedrock/privacy/templates/privacy/base-protocol.html
@@ -85,5 +85,6 @@
 {% block email_form %}{% endblock %}
 
 {% block js %}
+  {{ js_bundle('protocol-sidemenu') }}
   {{ js_bundle('privacy_protocol') }}
 {% endblock %}
