+++ bedrock/privacy/templates/privacy/notices/base-notice-paragraphs.html
@@ -39,5 +39,6 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('protocol-sidemenu') }}
   {{ js_bundle('privacy_protocol') }}
 {% endblock %}
