+++ bedrock/privacy/templates/privacy/notices/base-notice.html
@@ -33,5 +33,6 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('protocol-sidemenu') }}
   {{ js_bundle('privacy_protocol') }}
 {% endblock %}
