+++ media/js/base/protocol-globals.es6.js
@@ -1,0 +1,13 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+import MzpSupports from '@mozilla-protocol/core/protocol/js/supports';
+import MzpUtils from '@mozilla-protocol/core/protocol/js/utils';
+import MzpDetails from '@mozilla-protocol/core/protocol/js/details';
+
+window.MzpSupports = MzpSupports;
+window.MzpUtils = MzpUtils;
+window.MzpDetails = MzpDetails;
