+++ media/js/base/protocol/init-footer.es6.js
@@ -4,9 +4,9 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-(function () {
-    'use strict';
+/**
+ * Initialize Protocol footer.
+ */
+import MzpFooter from '@mozilla-protocol/core/protocol/js/footer';
 
-    // init menus on page
-    window.MzpDetails.init('.mzp-c-menu-list-title');
-})(window.Mzp);
+MzpFooter.init();
