+++ media/js/firefox/browsers-products.js
@@ -26,5 +26,5 @@
     }
 
     // init menus on page
-    window.Mzp.Details.init('.mzp-c-menu-list-title');
+    window.MzpDetails.init('.mzp-c-menu-list-title');
 })();
