+++ media/js/firefox/family/subnav.js
@@ -15,11 +15,11 @@
 
         // check we have global Supports and Details library
         if (
-            typeof Mzp.Supports !== 'undefined' &&
+            typeof window.MzpSupports !== 'undefined' &&
             typeof Mzp.Details !== 'undefined'
         ) {
             // check browser supports matchMedia
-            if (Mzp.Supports.matchMedia) {
+            if (window.MzpSupports.matchMedia) {
                 var _mqWide = matchMedia('(max-width: 1060px)');
 
                 // initialize details if screen is small
