+++ media/js/mozorg/about-transparency.js
@@ -8,13 +8,13 @@
     'use strict';
     var _mqWide = matchMedia('(max-width: 767px)');
     if (_mqWide.matches) {
-        window.Mzp.Details.init('.c-collapsible-section-heading');
+        window.MzpDetails.init('.c-collapsible-section-heading');
     }
     _mqWide.addListener(function (mq) {
         if (mq.matches) {
-            window.Mzp.Details.init('.c-collapsible-section-heading');
+            window.MzpDetails.init('.c-collapsible-section-heading');
         } else {
-            window.Mzp.Details.destroy('.c-collapsible-section-heading');
+            window.MzpDetails.destroy('.c-collapsible-section-heading');
         }
     });
 })();
