+++ media/js/newsletter/newsletter-init.es6.js
@@ -4,6 +4,8 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
+import MzpNewsletter from '@mozilla-protocol/core/protocol/js/newsletter';
 import NewsletterForm from './newsletter.es6';
 
+MzpNewsletter.init();
 NewsletterForm.init();
